# server/internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.KnownSitesResponse:
    properties:
      sites:
        items:
          type: string
        type: array
      user_content:
        items:
          type: string
        type: array
      version:
        type: string
    type: object
...
  handler.ThreatListResponse:
    properties:
      count:
        type: integer
      hash:
        example: sha256
        type: string
      prefix_len:
        example: 8
        type: integer
      prefixes:
        description: base64 of the sorted prefixes, back to back
        type: string
      sources:
        items:
          type: string
        type: array
      version:
        type: string
    type: object
...
info:
  contact: {}
  description: URL safety analysis API. Checks phishing signals, redirects, TLS, DNS,
    brand mismatches, and more.
  license:
    name: AGPL-3.0
  title: url.vet API
  version: "1.0"
paths:
...
  /lists/known-sites:
    get:
      description: The 10,000 most popular sites, and the user-content hosts on them
        that still need checking. A host matches a site if it is the site or a subdomain
        of it.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.KnownSitesResponse'
        "304":
          description: Not Modified
      summary: Well-known sites
      tags:
      - Lists
  /lists/threats:
    get:
      description: 'Known phishing and malware links (PhishTank, URLhaus) as sorted
        SHA-256 prefixes of each link''s normalized form: no scheme, lowercase host
        without "www.", no trailing slash or fragment, query kept. Clients hash a
        link the same way and look the prefix up locally, so nothing is sent. Supports
        If-None-Match.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ThreatListResponse'
        "304":
          description: Not Modified
      summary: Hashed threat list
      tags:
      - Lists
...
swagger: "2.0"
...